manager category

+++ models/categories-model.js
@@ -1,9 +1,23 @@
 const db=require('../utils/db');
+const countBookInCat= async(id)=>{
+    let count=await db.load(`SELECT count(*) as books from books where book_cat=${id}`);
+    return count[0].books;
+}
 module.exports={
     allCat:async()=>{
-        let cat=await db.load(`select cat_id as id, cat_name as name from categories`);
-        console.log(cat);
-        return cat;
+        let cats=await db.load(`select cat_id as id, cat_name as name from categories`);
+        console.log(cats);
+        return cats;
+    },
+    allCatAdmin:async()=>{
+        let cats=await db.load(`select cat_id as id, cat_name as name from categories`);
+        let i;
+        for(i in cats){
+            let c=await countBookInCat(cats[i].id);
+            cats[i].books=c;
+        }
+        console.log(cats);
+        return cats;
     },
     bookInCat:async(id)=>{
         let data=await db.load(`select b.book_id as id, b.book_name as name, c.cat_name as category, b.book_author as author, b.book_price as price
@@ -25,7 +39,9 @@ module.exports={
         return null;
     },
     deleteCat: async(id)=>{
-        db.load(`DELETE FROM categories WHERE cat_id =${id}`);
+        let count=await countBookInCat(id);
+        if(count!=0) return null;
+        await db.load(`DELETE FROM categories WHERE cat_id =${id}`);
         return null;
     },
